Show current user in Navbar and allow logout
