+++ apps/web/app/auth/signout/page.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { signOut, useSession } from "next-auth/react";
+import { useEffect } from "react";
+
+const SignOutPage = () => {
+  useEffect(() => {
+    signOut({ callbackUrl: "/" });
+  }, []);
+
+  return null;
+};
+
+export default SignOutPage;
