+++ apps/web/components/layout/NavBar/index.tsx
@@ -1,0 +1,452 @@
+"use client";
+
+import { Dialog, Menu, Transition } from "@headlessui/react";
+import { ChevronDownIcon } from "@heroicons/react/20/solid";
+import {
+  Bars3Icon,
+  BellIcon,
+  ChartPieIcon,
+  Cog6ToothIcon,
+  DocumentDuplicateIcon,
+  HomeIcon,
+  UsersIcon,
+  XMarkIcon,
+} from "@heroicons/react/24/outline";
+import { Session } from "next-auth";
+import Link from "next/link";
+import { usePathname } from "next/navigation";
+import { Fragment, useState } from "react";
+
+import { configuration } from "@/core/constants/configs";
+
+import { cn } from "@/lib/utils/common";
+
+const navigation = [
+  {
+    name: "Dashboard",
+    href: configuration.paths.dashbord,
+    icon: HomeIcon,
+  },
+  {
+    name: "Cap Table",
+    href: configuration.paths.capTables.all,
+    icon: UsersIcon,
+  },
+  {
+    name: "Share Class",
+    href: configuration.paths.shareClasses.all,
+    icon: DocumentDuplicateIcon,
+  },
+
+  {
+    name: "Equity Plan",
+    href: configuration.paths.equityPlans.all,
+    icon: ChartPieIcon,
+  },
+];
+const compliances = [
+  {
+    id: 1,
+    name: "409A",
+    href: configuration.paths.valuation401A,
+    initial: "4",
+    current: false,
+  },
+];
+const forms = [
+  {
+    id: 1,
+    name: "Rule 701",
+    href: configuration.paths.rule701,
+    initial: "R",
+    current: false,
+  },
+];
+const userNavigation = [
+  { name: "Profile", href: "/dashboard/profile" },
+  { name: "Sign out", href: "/auth/signout" },
+];
+
+export default function NavBar({ user }: { user: Session["user"] | null }) {
+  const pathname = usePathname();
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+
+  return (
+    <>
+      <Transition.Root show={sidebarOpen} as={Fragment}>
+        <Dialog
+          as="div"
+          className="relative z-50 lg:hidden"
+          onClose={setSidebarOpen}
+        >
+          <Transition.Child
+            as={Fragment}
+            enter="transition-opacity ease-linear duration-300"
+            enterFrom="opacity-0"
+            enterTo="opacity-100"
+            leave="transition-opacity ease-linear duration-300"
+            leaveFrom="opacity-100"
+            leaveTo="opacity-0"
+          >
+            <div className="fixed inset-0 bg-gray-900/80" />
+          </Transition.Child>
+
+          <div className="fixed inset-0 flex">
+            <Transition.Child
+              as={Fragment}
+              enter="transition ease-in-out duration-300 transform"
+              enterFrom="-translate-x-full"
+              enterTo="translate-x-0"
+              leave="transition ease-in-out duration-300 transform"
+              leaveFrom="translate-x-0"
+              leaveTo="-translate-x-full"
+            >
+              <Dialog.Panel className="relative mr-16 flex w-full max-w-xs flex-1">
+                <Transition.Child
+                  as={Fragment}
+                  enter="ease-in-out duration-300"
+                  enterFrom="opacity-0"
+                  enterTo="opacity-100"
+                  leave="ease-in-out duration-300"
+                  leaveFrom="opacity-100"
+                  leaveTo="opacity-0"
+                >
+                  <div className="absolute left-full top-0 flex w-16 justify-center pt-5">
+                    <button
+                      type="button"
+                      className="-m-2.5 p-2.5"
+                      onClick={() => setSidebarOpen(false)}
+                    >
+                      <span className="sr-only">Close sidebar</span>
+                      <XMarkIcon
+                        className="h-6 w-6 text-white"
+                        aria-hidden="true"
+                      />
+                    </button>
+                  </div>
+                </Transition.Child>
+                {/* Sidebar component, swap this element with another sidebar if you like */}
+                <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-white px-6 pb-4">
+                  <div className="flex h-16 shrink-0 items-center">
+                    cap.octolane.com
+                  </div>
+                  <nav className="flex flex-1 flex-col">
+                    <ul role="list" className="flex flex-1 flex-col gap-y-7">
+                      <li>
+                        <ul role="list" className="-mx-2 space-y-1">
+                          {navigation.map(item => {
+                            const isCurrent = pathname.startsWith(item.href);
+
+                            return (
+                              <li key={item.name}>
+                                <Link
+                                  href={item.href}
+                                  className={cn(
+                                    isCurrent
+                                      ? "bg-gray-50 text-zinc-600"
+                                      : "text-gray-700 hover:text-zinc-600 hover:bg-gray-50",
+                                    "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold",
+                                  )}
+                                >
+                                  <item.icon
+                                    className={cn(
+                                      isCurrent
+                                        ? "text-zinc-600"
+                                        : "text-gray-400 group-hover:text-zinc-600",
+                                      "h-6 w-6 shrink-0",
+                                    )}
+                                    aria-hidden="true"
+                                  />
+                                  {item.name}
+                                </Link>
+                              </li>
+                            );
+                          })}
+                        </ul>
+                      </li>
+                      <li>
+                        <div className="text-xs font-semibold leading-6 text-gray-400">
+                          Comply & Tax
+                        </div>
+                        <ul role="list" className="-mx-2 mt-2 space-y-1">
+                          {compliances.map(comply => {
+                            const isCurrent = pathname.startsWith(comply.href);
+                            return (
+                              <li key={comply.name}>
+                                <Link
+                                  href={comply.href}
+                                  className={cn(
+                                    isCurrent
+                                      ? "bg-gray-50 text-zinc-600"
+                                      : "text-gray-700 hover:text-zinc-600 hover:bg-gray-50",
+                                    "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold",
+                                  )}
+                                >
+                                  <span
+                                    className={cn(
+                                      isCurrent
+                                        ? "text-zinc-600 border-zinc-600"
+                                        : "text-gray-400 border-gray-200 group-hover:border-zinc-600 group-hover:text-zinc-600",
+                                      "flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border text-[0.625rem] font-medium bg-white",
+                                    )}
+                                  >
+                                    {comply.initial}
+                                  </span>
+                                  <span className="truncate">
+                                    {comply.name}
+                                  </span>
+                                </Link>
+                              </li>
+                            );
+                          })}
+                        </ul>
+                      </li>
+                      <li className="mt-auto">
+                        <Link
+                          href={configuration.paths.settings}
+                          className="group -mx-2 flex gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-700 hover:bg-gray-50 hover:text-zinc-600"
+                        >
+                          <Cog6ToothIcon
+                            className="h-6 w-6 shrink-0 text-gray-400 group-hover:text-zinc-600"
+                            aria-hidden="true"
+                          />
+                          Settings
+                        </Link>
+                      </li>
+                    </ul>
+                  </nav>
+                </div>
+              </Dialog.Panel>
+            </Transition.Child>
+          </div>
+        </Dialog>
+      </Transition.Root>
+
+      {/* Static sidebar for desktop */}
+      <div className="hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col">
+        {/* Sidebar component, swap this element with another sidebar if you like */}
+        <div className="flex grow flex-col gap-y-5 overflow-y-auto border-r border-gray-200 bg-white px-6 pb-4">
+          <div className="flex h-16 shrink-0 items-center">
+            cap.octolane.com
+          </div>
+          <nav className="flex flex-1 flex-col">
+            <ul role="list" className="flex flex-1 flex-col gap-y-7">
+              <li>
+                <ul role="list" className="-mx-2 space-y-1">
+                  {navigation.map(item => {
+                    const isCurrent = pathname.startsWith(item.href);
+                    return (
+                      <li key={item.name}>
+                        <Link
+                          href={item.href}
+                          className={cn(
+                            isCurrent
+                              ? "bg-gray-50 text-zinc-600"
+                              : "text-gray-700 hover:text-zinc-600 hover:bg-gray-50",
+                            "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold",
+                          )}
+                        >
+                          <item.icon
+                            className={cn(
+                              isCurrent
+                                ? "text-zinc-600"
+                                : "text-gray-400 group-hover:text-zinc-600",
+                              "h-6 w-6 shrink-0",
+                            )}
+                            aria-hidden="true"
+                          />
+                          {item.name}
+                        </Link>
+                      </li>
+                    );
+                  })}
+                </ul>
+              </li>
+              <li>
+                <div className="text-xs font-semibold leading-6 text-gray-400">
+                  Valuations
+                </div>
+                <ul role="list" className="-mx-2 mt-2 space-y-1">
+                  {compliances.map(team => {
+                    const isCurrent = pathname.startsWith(team.href);
+
+                    return (
+                      <li key={team.name}>
+                        <Link
+                          href={team.href}
+                          className={cn(
+                            isCurrent
+                              ? "bg-gray-50 text-zinc-600"
+                              : "text-gray-700 hover:text-zinc-600 hover:bg-gray-50",
+                            "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold",
+                          )}
+                        >
+                          <span
+                            className={cn(
+                              isCurrent
+                                ? "text-zinc-600 border-zinc-600"
+                                : "text-gray-400 border-gray-200 group-hover:border-zinc-600 group-hover:text-zinc-600",
+                              "flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border text-[0.625rem] font-medium bg-white",
+                            )}
+                          >
+                            {team.initial}
+                          </span>
+                          <span className="truncate">{team.name}</span>
+                        </Link>
+                      </li>
+                    );
+                  })}
+                </ul>
+              </li>
+              <li>
+                <div className="text-xs font-semibold leading-6 text-gray-400">
+                  IRS & SEC Forms
+                </div>
+                <ul role="list" className="-mx-2 mt-2 space-y-1">
+                  {forms.map(form => {
+                    const isCurrent = pathname.startsWith(form.href);
+                    return (
+                      <li key={form.name}>
+                        <Link
+                          href={form.href}
+                          className={cn(
+                            isCurrent
+                              ? "bg-gray-50 text-zinc-600"
+                              : "text-gray-700 hover:text-zinc-600 hover:bg-gray-50",
+                            "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold",
+                          )}
+                        >
+                          <span
+                            className={cn(
+                              isCurrent
+                                ? "text-zinc-600 border-zinc-600"
+                                : "text-gray-400 border-gray-200 group-hover:border-zinc-600 group-hover:text-zinc-600",
+                              "flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border text-[0.625rem] font-medium bg-white",
+                            )}
+                          >
+                            {form.initial}
+                          </span>
+                          <span className="truncate">{form.name}</span>
+                        </Link>
+                      </li>
+                    );
+                  })}
+                </ul>
+              </li>
+              <li className="mt-auto">
+                <Link
+                  href={configuration.paths.settings}
+                  className="group -mx-2 flex gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-700 hover:bg-gray-50 hover:text-zinc-600"
+                >
+                  <Cog6ToothIcon
+                    className="h-6 w-6 shrink-0 text-gray-400 group-hover:text-zinc-600"
+                    aria-hidden="true"
+                  />
+                  Settings
+                </Link>
+              </li>
+            </ul>
+          </nav>
+        </div>
+      </div>
+
+      <div className="lg:pl-72">
+        <div className="sticky top-0 z-40 lg:mx-auto lg:max-w-7xl lg:px-8">
+          <div className="flex h-16 items-center gap-x-4 border-b border-gray-200 bg-white px-4 shadow-sm sm:gap-x-6 sm:px-6 lg:px-0 lg:shadow-none">
+            <button
+              type="button"
+              className="-m-2.5 p-2.5 text-gray-700 lg:hidden"
+              onClick={() => setSidebarOpen(true)}
+            >
+              <span className="sr-only">Open sidebar</span>
+              <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+            </button>
+
+            {/* Separator */}
+            <div
+              className="h-6 w-px bg-gray-200 lg:hidden"
+              aria-hidden="true"
+            />
+
+            <div className="flex flex-1 gap-x-4 self-stretch lg:gap-x-6">
+              <form
+                className="relative flex flex-1"
+                action="#"
+                method="GET"
+              ></form>
+              <div className="flex items-center gap-x-4 lg:gap-x-6">
+                <button
+                  type="button"
+                  className="-m-2.5 p-2.5 text-gray-400 hover:text-gray-500"
+                >
+                  <span className="sr-only">View notifications</span>
+                  <BellIcon className="h-6 w-6" aria-hidden="true" />
+                </button>
+
+                {/* Separator */}
+                <div
+                  className="hidden lg:block lg:h-6 lg:w-px lg:bg-gray-200"
+                  aria-hidden="true"
+                />
+
+                {/* Profile dropdown */}
+                <Menu as="div" className="relative">
+                  <Menu.Button className="-m-1.5 flex items-center p-1.5">
+                    <span className="sr-only">Open user menu</span>
+                    <img
+                      className="h-8 w-8 rounded-full bg-gray-50"
+                      src={
+                        user?.image ||
+                        `https://ui-avatars.com/api/?name=${user?.name}`
+                      }
+                      alt=""
+                    />
+                    <span className="hidden lg:flex lg:items-center">
+                      <span
+                        className="ml-4 text-sm font-semibold leading-6 text-gray-900"
+                        aria-hidden="true"
+                      >
+                        {user?.name}
+                      </span>
+                      <ChevronDownIcon
+                        className="ml-2 h-5 w-5 text-gray-400"
+                        aria-hidden="true"
+                      />
+                    </span>
+                  </Menu.Button>
+                  <Transition
+                    as={Fragment}
+                    enter="transition ease-out duration-100"
+                    enterFrom="transform opacity-0 scale-95"
+                    enterTo="transform opacity-100 scale-100"
+                    leave="transition ease-in duration-75"
+                    leaveFrom="transform opacity-100 scale-100"
+                    leaveTo="transform opacity-0 scale-95"
+                  >
+                    <Menu.Items className="absolute right-0 z-10 mt-2.5 w-32 origin-top-right rounded-md bg-white py-2 shadow-lg ring-1 ring-gray-900/5 focus:outline-none">
+                      {userNavigation.map(item => (
+                        <Menu.Item key={item.name}>
+                          {({ active }) => (
+                            <Link
+                              href={item.href}
+                              className={cn(
+                                active ? "bg-gray-50" : "",
+                                "block px-3 py-1 text-sm leading-6 text-gray-900",
+                              )}
+                            >
+                              {item.name}
+                            </Link>
+                          )}
+                        </Menu.Item>
+                      ))}
+                    </Menu.Items>
+                  </Transition>
+                </Menu>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
